cleaned up a bit of codebase, moved many methods from parse.py to stats.py to better represent the files' names
